move low-level reproject stuff into separate file
